+++ nilearn/plotting/tests/test_img_plotting.py
@@ -7,6 +7,7 @@ from distutils.version import LooseVersion
 
 import matplotlib
 import matplotlib.pyplot as plt
+from matplotlib.patches import FancyArrow
 import nibabel
 import numpy as np
 import pytest
@@ -60,10 +61,21 @@ def testdata_4d():
         rng.uniform(size=(7, 7, 3, 1777)), mni_affine
     )
     img_mask = nibabel.Nifti1Image(np.ones((7, 7, 3), int), mni_affine)
+    atlas = np.ones((7, 7, 3), int)
+    atlas[2:5, :, :] = 2
+    atlas[5:8, :, :] = 3
+    img_atlas = nibabel.Nifti1Image(atlas, mni_affine)
+    atlas_labels = {
+        "gm": 1,
+        "wm": 2,
+        "csf": 3,
+    }
     data = {
         'img_4d': img_4d,
         'img_4d_long': img_4d_long,
         'img_mask': img_mask,
+        'img_atlas': img_atlas,
+        'atlas_labels': atlas_labels,
     }
     return data
 
@@ -173,6 +185,19 @@ def test_plot_functions(testdata_3d, testdata_4d, tmpdir):
     plot_stat_map(img_3d, symmetric_cbar=True,
                   output_file=filename,
                   axes=ax, vmax=np.nan)
+    plt.close()
+
+
+def test_plot_functions_mosaic_mode(testdata_3d):
+    img_3d = testdata_3d['img']
+
+    # smoke-test for 3D plotting functions with display_mode='mosaic'
+    for cut_coords in [None, 5, (5, 4, 3)]:
+        for plot_func in [plot_anat, plot_img, plot_stat_map, plot_epi,
+                          plot_roi]:
+            plot_func(img_3d, display_mode='mosaic',
+                      title='mosaic mode', cut_coords=cut_coords)
+
     plt.close()
 
 
@@ -300,8 +325,7 @@ def test_plot_glass_brain_threshold_for_uint8(testdata_3d):
 
 
 def test_plot_carpet(testdata_4d):
-    """Check contents of plot_carpet figure against data in image.
-    """
+    """Check contents of plot_carpet figure against data in image."""
     img_4d = testdata_4d['img_4d']
     img_4d_long = testdata_4d['img_4d_long']
     mask_img = testdata_4d['img_mask']
@@ -328,6 +352,60 @@ def test_plot_carpet(testdata_4d):
     # Check size
     n_items = (np.prod(img_4d_long.shape[:-1]) * np.ceil(img_4d_long.shape[-1] / 4))
     assert plotted_array.size == n_items
+    plt.close(display)
+
+
+def test_plot_carpet_with_atlas(testdata_4d):
+    """Test plot_carpet when using an atlas."""
+    img_4d = testdata_4d['img_4d']
+    mask_img = testdata_4d['img_atlas']
+    atlas_labels = testdata_4d['atlas_labels']
+
+    # Test atlas - labels
+    display = plot_carpet(img_4d, mask_img, detrend=False, title='TEST')
+
+    # Check the output
+    # Two axes: 1 for colorbar and 1 for imshow
+    assert len(display.axes) == 2
+    # The y-axis label of the imshow should be 'voxels' since atlas labels are
+    # unknown
+    ax = display.axes[1]
+    assert ax.get_ylabel() == 'voxels'
+
+    # Next two lines retrieve the numpy array from the plot
+    ax = display.axes[0]
+    colorbar = ax.images[0].get_array()
+    assert len(np.unique(colorbar)) == len(atlas_labels)
+
+    # Save execution time and memory
+    plt.close(display)
+
+    # Test atlas + labels
+    fig, ax = plt.subplots()
+    display = plot_carpet(
+        img_4d,
+        mask_img,
+        mask_labels=atlas_labels,
+        detrend=True,
+        title='TEST',
+        figure=fig,
+        axes=ax,
+    )
+    # Check the output
+    # Two axes: 1 for colorbar and 1 for imshow
+    assert len(display.axes) == 2
+    ax = display.axes[0]
+
+    # The ytick labels of the colorbar should match the atlas labels
+    yticklabels = ax.get_yticklabels()
+    yticklabels = [yt.get_text() for yt in yticklabels]
+    assert set(yticklabels) == set(atlas_labels.keys())
+
+    # Next two lines retrieve the numpy array from the plot
+    ax = display.axes[0]
+    colorbar = ax.images[0].get_array()
+    assert len(np.unique(colorbar)) == len(atlas_labels)
+
     plt.close(display)
 
 
@@ -566,6 +644,58 @@ def test_plot_connectome(tmpdir):
     plot_connectome(*args, node_color=['red'], display_mode='lzry')
     plt.close()
 
+    # Non symmetric matrix
+    adjacency_matrix = np.array([[1., -2., 0.3, 0.2],
+                                 [0.1, 1, 1.1, 0.1],
+                                 [0.01, 2.3, 1., 3.1],
+                                 [0.6, 0.03, 1.2, 1.]])
+    ax = plot_connectome(adjacency_matrix,
+                         node_coords,
+                         display_mode='ortho')
+    # No thresholding was performed, we should get
+    # as many arrows as we have edges
+    for direction in ['x', 'y', 'z']:
+        assert(len([patch for patch in ax.axes[direction].ax.patches
+             if isinstance(patch, FancyArrow)]) ==
+                    np.prod(adjacency_matrix.shape))
+
+    # Set a few elements of adjacency matrix to zero
+    adjacency_matrix[1, 0] = 0.0
+    adjacency_matrix[2, 3] = 0.0
+    # Plot with different display mode
+    ax = plot_connectome(adjacency_matrix,
+                         node_coords,
+                         display_mode='lzry')
+    # No edge in direction 'l' because of node coords
+    assert(len([patch for patch in ax.axes['l'].ax.patches
+             if isinstance(patch, FancyArrow)]) == 0)
+    for direction in ['z', 'r', 'y']:
+        assert(len([patch for patch in ax.axes[direction].ax.patches
+             if isinstance(patch, FancyArrow)]) ==
+                    np.prod(adjacency_matrix.shape) - 2)
+
+    # Edge thresholding
+    # Case 1: Threshold is a number
+    thresh = 1.1
+    ax = plot_connectome(adjacency_matrix,
+                         node_coords,
+                         edge_threshold=thresh)
+    for direction in ['x', 'y', 'z']:
+        assert(len([patch for patch in ax.axes[direction].ax.patches
+             if isinstance(patch, FancyArrow)]) ==
+                    np.sum(np.abs(adjacency_matrix) >= thresh))
+    # Case 2: Threshold is a percentage
+    thresh = 80
+    ax = plot_connectome(adjacency_matrix,
+                         node_coords,
+                         edge_threshold="{}%".format(thresh))
+    for direction in ['x', 'y', 'z']:
+        assert(len([patch for patch in ax.axes[direction].ax.patches
+             if isinstance(patch, FancyArrow)]) ==
+               np.sum(np.abs(adjacency_matrix) >=
+                    np.percentile(np.abs(
+                        adjacency_matrix.ravel()), thresh)))
+
 
 def test_plot_connectome_exceptions():
     node_coords = np.arange(2 * 3).reshape((2, 3))
@@ -577,7 +707,9 @@ def test_plot_connectome_exceptions():
     # adjacency_matrix is not symmetric
     non_symmetric_adjacency_matrix = np.array([[1., 2],
                                                [0.4, 1.]])
-    with pytest.raises(ValueError, match='should be symmetric'):
+    with pytest.warns(UserWarning,
+                      match=("'adjacency_matrix' is not symmetric. "
+                             "A directed graph will be plotted.")):
         plot_connectome(non_symmetric_adjacency_matrix, node_coords, **kwargs)
 
     adjacency_matrix = np.array([[1., 2.],
@@ -586,7 +718,10 @@ def test_plot_connectome_exceptions():
     masked_adjacency_matrix = np.ma.masked_array(
         adjacency_matrix, [[False, True], [False, False]])
 
-    with pytest.raises(ValueError, match='non symmetric mask'):
+    with pytest.warns(UserWarning,
+                      match=("'adjacency_matrix' was masked with "
+                             "a non symmetric mask. A directed "
+                             "graph will be plotted.")):
         plot_connectome(masked_adjacency_matrix, node_coords, **kwargs)
 
     # edges threshold is neither a number nor a string
@@ -1001,6 +1136,18 @@ def test_invalid_in_display_mode_tiled_cut_coords_all_plots(testdata_3d):
                                  "match the display_mode"
                            ):
             plot_func(img, display_mode='tiled', cut_coords=(2, 2))
+
+
+def test_invalid_in_display_mode_mosaic_cut_coords_all_plots(testdata_3d):
+    img = testdata_3d['img']
+
+    for plot_func in [plot_img, plot_anat, plot_roi, plot_epi,
+                      plot_stat_map, plot_prob_atlas]:
+        with pytest.raises(ValueError,
+                           match="The number cut_coords passed does not "
+                                 "match the display_mode"
+                           ):
+            plot_func(img, display_mode='mosaic', cut_coords=(2, 2))
 
 
 def test_outlier_cut_coords():